third_party/WebKit/LayoutTests/editing/pasteboard/cut-text-001.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable>foo ^bar| baz</div>',
        selection => {
            assert_not_equals(window.internals, undefined,
                'This test requires window.internals to access clipboard');
            selection.document.execCommand('cut');
        },
        '<div contenteditable>foo\u{00A0}| baz</div>'),
        'cut command');
</script>
